merge_upload_bot/requirements.txt
pyrogram
tgcrypto
ffmpeg-python
pymediainfo
aiofiles
aiogram==3.4
python-dotenv
ffmpeg-python
pymediainfo
pytz
requests
rich
tqdm
python-magic
uvloop
watchdog
